card-service/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
Flask==2.2.2
gunicorn==20.1.0
requests==2.31.0
pymongo==4.3.3
python-dotenv==1.0.0